Move findBreakpoint to its own file
